added device-ui to lightSleep observers for handling graceful sleep; refactoring main.cpp
